Добавление менеджера драйверов, фикс сюиты и очистки стораджа

+++ src/test/java/tests/base/BaseTest.java
@@ -1,8 +1,10 @@
 package tests.base;
 
 import common.CommonActions;
+import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
+import org.testng.annotations.BeforeSuite;
 import pages.base.BasePage;
 import org.testng.annotations.AfterTest;
 import org.testng.annotations.AfterSuite;
@@ -21,13 +23,14 @@ public class BaseTest {
     protected PayDocBy payDocBy = new PayDocBy(driver);
 
 
-
     @AfterTest
     public void clearCookiesAndLocalStorage (){
         if (CLEAR_COOKIES_AND_STORAGE){
             JavascriptExecutor javascriptExecutor = (JavascriptExecutor) driver;
             driver.manage().deleteAllCookies();
-            javascriptExecutor.executeScript("window.sessionStorage.clear()");
+            if(!driver.getCurrentUrl().contains("data:")){
+                javascriptExecutor.executeScript("window.sessionStorage.clear()");
+            }
         }
 
     }
